Undo splitting state
