+++ src/parsita/parsers/_repeated_seperated.py
@@ -2,8 +2,7 @@ __all__ = ["RepeatedSeparatedParser", "repsep", "RepeatedOnceSeparatedParser", "
 
 from typing import Any, Generic, Optional, Sequence, Union
 
-from ..exceptions import RecursionError
-from ..state import Continue, Input, Output, Reader, State
+from ..state import Continue, Input, Output, Reader, RecursionError, State
 from ._base import Parser
 from ._literal import lit
